refactor out the help system into its own component - use more accurate event names for opening/closing tooltips - moved the help system code into the old help component - removed the old unused menu component
