+++ src/client/components/editor/index.js
@@ -1,14 +1,12 @@
 const React = require('react');
 const ReactDom = require('react-dom');
 const h = require('react-hyperscript');
-const hh = require('hyperscript');
-const tippyjs = require('tippy.js');
 const EventEmitter = require('eventemitter3');
 const io = require('socket.io-client');
 const _ = require('lodash');
 const Promise = require('bluebird');
 
-const { getId, defer, makeClassList, isInteractionNode } = require('../../../util');
+const { getId, defer } = require('../../../util');
 const Document = require('../../../model/document');
 
 const Notification = require('../notification');
@@ -16,12 +14,12 @@ const CornerNotification = require('../notification/corner');
 
 const logger = require('../../logger');
 const debug = require('../../debug');
-const { tippyDefaults, tippyTopZIndex } = require('../../defs');
 
 const makeCytoscape = require('./cy');
 const defs = require('./defs');
 const Buttons = require('./buttons');
 const UndoRemove = require('./undo-remove');
+const Help = require('./help');
 
 const RM_DEBOUNCE_TIME = 500;
 const RM_AVAIL_DURATION = 5000;
@@ -117,7 +115,6 @@ class Editor extends React.Component {
     this.data = ({
       bus: bus,
       document: doc,
-      showHelp: false,
       drawMode: false,
       newElementShift: 0,
       mountDeferred: defer(),
@@ -333,22 +330,17 @@ class Editor extends React.Component {
   }
 
   render(){
-    let { document, bus, incompleteNotification } = this.data;
+    let { document, bus, incompleteNotification, cy } = this.data;
     let controller = this;
 
     let editorContent = this.state.initted ? [
       h(Buttons, { controller, document, bus }),
-      incompleteNotification  && !this.state.showHelp ? h(CornerNotification, { notification: incompleteNotification }) : h('span'),
+      incompleteNotification ? h(CornerNotification, { notification: incompleteNotification }) : h('span'),
       h(UndoRemove, { controller, document, bus }),
       h('div.editor-graph#editor-graph'),
-      h('div.editor-help-overlay', { onClick: () => this.toggleHelp(), className: makeClassList({'editor-help-overlay-active': this.state.showHelp})} )
+      h(Help, { bus, cy })
     ] : [];
 
-    if( this.state.showHelp ){
-       editorContent.push(h('div.editor-button.editor-help-close-button', { onClick: () => this.toggleHelp() }, [
-        h('i.material-icons', 'close')
-       ]));
-    }
     return h('div.editor' + ( this.state.initted ? '.editor-initted' : '' ), editorContent);
   }
 
@@ -364,71 +356,6 @@ class Editor extends React.Component {
     }
   }
 
-  toggleHelp(){
-    let showHelp = this.data.showHelp;
-    let bus = this.data.bus;
-    let cy = this.data.cy;
-
-    if( !showHelp ){
-      bus.emit('showhelp');
-
-      let ent = cy.nodes(node => !isInteractionNode(node)).first();
-      if( !ent.empty() ){
-        let entRef = ent.popperRef();
-        let entTippy = new tippyjs(entRef, _.assign({}, tippyDefaults, {
-          html: (() => {
-            return hh('div.editor-help-tooltip', [
-              hh('div', 'Provide entity name'),
-              hh('ul', [
-                hh('li', 'E.g P53')
-              ])
-            ]);
-          })(),
-          theme: 'dark',
-          placement: 'right',
-          trigger: 'manual',
-          distance: 32,
-          zIndex: tippyTopZIndex
-        })).tooltips[0];
-        this.data.entTip = entTippy;
-        entTippy.show();
-      }
-
-      let intn = cy.nodes(isInteractionNode).first();
-      if( !intn.empty() ){
-        let intnRef = intn.popperRef();
-        let intnTippy = new tippyjs(intnRef, _.assign({}, tippyDefaults, {
-          html: (() => {
-            return hh('div.editor-help-tooltip', [
-              hh('div', 'Provide interaction type and direction'),
-              hh('ul', [
-                hh('li', 'E.g A activates phosphorylation of B')
-              ])
-            ]);
-          })(),
-          theme: 'dark',
-          trigger: 'manual',
-          placement: 'left',
-          distance: 32,
-          zIndex: tippyTopZIndex
-        })).tooltips[0];
-
-        this.data.intnTip = intnTippy;
-        intnTippy.show();
-      }
-      this.setData({ showHelp: true });
-    } else {
-      bus.emit('closehelp');
-      if( this.data.entTip ){
-        this.data.entTip.hide();
-      }
-      if( this.data.intnTip ){
-        this.data.intnTip.hide();
-      }
-      this.setData({ showHelp: false });
-    }
-  }
-
   componentWillUnmount(){
     let { cy, document, bus } = this.data;
 
@@ -436,12 +363,6 @@ class Editor extends React.Component {
 
     if( cy ){
       cy.destroy();
-    }
-    if( this.data.entTip ){
-      this.data.entTip.destroy();
-    }
-    if( this.data.intnTip ){
-      this.data.intnTip.destroy();
     }
 
     document.elements().forEach( el => el.removeAllListeners() );
